# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API endpoint",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/company": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get list of Companies",
                "description": "Fetches a list of all companies available.",
                "operationId": "getcompanies",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Create a company",
                "operationId": "addCompany",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanySuccesResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/company/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get company",
                "description": "Fetches information about a specific company by its ID",
                "operationId": "getCompanyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanySuccesResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Update specific company",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "company ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanySuccesResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Delete specific Company",
                "operationId": "deleteCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "company ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "it is deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoContentResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/employee": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get list of Employees",
                "description": "Fetches a list of all employees available.",
                "operationId": "getemployees",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Create a employee",
                "operationId": "addEmployee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeSuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/employee/{id}": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get employee",
                "description": "Fetches information about a specific employee by its ID",
                "operationId": "getEmployeeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Employee"
                ],
                "summary": "Update specific employee",
                "operationId": "updateEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeSuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Delete specific Employee",
                "operationId": "deleteEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "it is deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoContentResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CompanyRequestBody": {
                "required": [
                    "companyName",
                    "location"
                ],
                "properties": {
                    "companyName": {
                        "type": "string",
                        "example": "Google"
                    },
                    "location": {
                        "type": "string",
                        "example": "Moscow"
                    }
                },
                "type": "object"
            },
            "EmployeeRequestBody": {
                "required": [
                    "name",
                    "position",
                    "companyId"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John"
                    },
                    "position": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "companyId": {
                        "type": "string",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "CompanyListResponse": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CompanySchema"
                        }
                    }
                },
                "type": "object"
            },
            "CompanySuccesResponse": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/CompanySchema"
                    }
                },
                "type": "object"
            },
            "EmployeeListResponse": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EmployeeSchema"
                        }
                    }
                },
                "type": "object"
            },
            "EmployeeSuccessResponse": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/EmployeeSchema"
                    }
                },
                "type": "object"
            },
            "NoContentResponse": {
                "description": "No Content",
                "properties": {
                    "data": {
                        "description": "No Content",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
            "UnprocessableEntityResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unprocessable Content"
                    }
                },
                "type": "object"
            },
            "CompanySchema": {
                "required": [
                    "companyName",
                    "location"
                ],
                "properties": {
                    "companyName": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Google"
                    },
                    "location": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Moscow"
                    }
                },
                "type": "object"
            },
            "EmployeeSchema": {
                "required": [
                    "name",
                    "position",
                    "companyId"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "John"
                    },
                    "position": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "admin"
                    },
                    "companyId": {
                        "type": "string",
                        "maxLength": 64,
                        "example": 2
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "CompanyRequestBody": {
                "description": "company request body",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CompanyRequestBody"
                        }
                    }
                }
            },
            "EmployeeRequestBody": {
                "description": "employee request body",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/EmployeeRequestBody"
                        }
                    }
                }
            }
        }
    }
}